Rettet opp feil i parsing og tabell-oppdateringer

+++ src/main/kotlin/no/nav/sokos/spk.mottak/modell/EndRecord.kt
@@ -4,5 +4,5 @@ import java.math.BigDecimal
 
 data class EndRecord(
     val numberOfRecord: Int,
-    val totalBelop: BigDecimal
+    val totalBelop: Long
 )
